+++ server/component/style/conditionalContainer/ConditionalContainerView.php
@@ -27,6 +27,11 @@ class ConditionalContainerView extends StyleView
      */
     private $debug;
 
+    /**
+     * For which platform the condition should be checked, by default the style checks for both platforms
+     */
+    private $platform;
+
     /* Constructors ***********************************************************/
 
     /**
@@ -40,12 +45,29 @@ class ConditionalContainerView extends StyleView
         parent::__construct($model);
         $this->condition = $this->model->get_db_field('condition');
         $this->debug = $this->model->get_db_field('debug', false);
+        $this->platform = $this->model->get_db_field('platform', pageAccessTypes_mobile_and_web);
     }
 
     private function get_entry_values($entry_value){
         $cond = json_encode($this->model->get_db_field('condition'));
         $cond = $this->model->get_entry_value($entry_value, $cond);
         $this->condition = json_decode($cond, true);
+    }
+
+    /**
+     * Check if the call is from the platform that was selected for the style
+     * @retval boolean
+     * Returns true if the call is from the correct platform otherwise false
+     */
+    private function is_correct_platform(){
+        if ($this->platform == pageAccessTypes_mobile_and_web) {
+            return true;
+        } else if ($this->platform == pageAccessTypes_mobile && (isset($_POST['mobile']) && $_POST['mobile'])) {
+            return true;
+        } else if ($this->platform == pageAccessTypes_web && !isset($_POST['mobile'])) {
+            return true;
+        }
+        return false;
     }
 
     /* Public Methods *********************************************************/
@@ -55,6 +77,10 @@ class ConditionalContainerView extends StyleView
      */
     public function output_content()
     {
+        if(!$this->is_correct_platform()){
+            // break it is not the correct platform
+            return;
+        }
         $res = $this->model->compute_condition($this->condition);
         if($this->debug)
         {
@@ -73,6 +99,10 @@ class ConditionalContainerView extends StyleView
      */
     public function output_content_entry($entry_value)
     {
+        if(!$this->is_correct_platform()){
+            // break it is not the correct platform
+            return;
+        }
         $entry_data = $entry_value;
         $this->get_entry_values($entry_value);
 
@@ -89,6 +119,10 @@ class ConditionalContainerView extends StyleView
 
     public function output_content_mobile()
     {
+        if(!$this->is_correct_platform()){
+            // break it is not the correct platform
+            return [];
+        }
         $res = $this->model->compute_condition($this->condition);
         if ($res['result']) {
             return parent::output_content_mobile();
@@ -102,6 +136,10 @@ class ConditionalContainerView extends StyleView
      */
     public function output_content_mobile_entry($entry_value)
     {        
+        if(!$this->is_correct_platform()){
+            // break it is not the correct platform
+            return [];
+        }
         $this->get_entry_values($entry_value);
         $res = $this->model->compute_condition($this->condition);
         if ($res['result']) {