+++ server/db/update_v4.0.0_v4.1.0.sql
@@ -91,4 +91,10 @@ where id_fields = get_field_id('ajax') and id_styles = get_style_id('formUserInp
 
 # executed on studybuddy
 
+-- Add new field type `select-platform` and field `platform` in style conditionalCointainer
+INSERT INTO `fieldType` (`id`, `name`, `position`) VALUES (NULL, 'select-platform', '8');
+INSERT INTO `fields` (`id`, `name`, `id_type`, `display`) VALUES (NULL, 'platform', get_field_type_id('select-platform'), '0');
+INSERT INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) 
+VALUES (get_style_id('conditionalContainer'), get_field_id('platform'), 'mobile_and_web', 'Select for which platform the conditional container will be loaded');
+
 #gen all functions and procs